Introduce the new Sourcepawn Condenser
I introduce you: the new sourcepawn condenser. A small Lexer based on a
simple tokenizer. Nevertheless, it's fast enough to provide Informations
on the fly and enables us, to parse files which are actually worked on
and therefore adds the possibility to a dynamic intellisense.
It's btw. extremly fast compared to the old regex condenser. It parses
the whole sourcemod include+smlib+other libraries in less than 20
milliseconds.